patient can be added to session created w/out patient, fixed reviewing public session results, review session when adding patient to session

+++ app/src/main/java/com/example/rafael/appprototype/SharedPreferencesHelper.java
@@ -14,6 +14,8 @@ import static android.content.Context.MODE_PRIVATE;
 
 public class SharedPreferencesHelper {
 
+    private static boolean loggedIn;
+
     public static String isThereOngoingPublicSession(Context context) {
         SharedPreferences sharedPreferences = context.getSharedPreferences(context.getString(R.string.sharedPreferencesTag), MODE_PRIVATE);
         return sharedPreferences.getString(context.getResources().getString(R.string.saved_session_public), null);
@@ -25,11 +27,19 @@ public class SharedPreferencesHelper {
     }
 
     public static void resetPublicSession(Activity context, String sessionID) {
-        Session session = Session.getSessionByID(sessionID);
-        if (session != null)
-            session.delete();
+        if (sessionID != null) {
+            Session session = Session.getSessionByID(sessionID);
+            if (session != null)
+                session.delete();
+        }
+
         SharedPreferences sharedPreferences = context.getSharedPreferences(context.getString(R.string.sharedPreferencesTag), MODE_PRIVATE);
         sharedPreferences.edit().putString(context.getString(R.string.saved_session_public), null).apply();
+    }
+
+    public static void setPrivateSession(Activity context, String sessionID) {
+        SharedPreferences sharedPreferences = context.getSharedPreferences(context.getString(R.string.sharedPreferencesTag), MODE_PRIVATE);
+        sharedPreferences.edit().putString(context.getString(R.string.saved_session_private), sessionID).apply();
     }
 
     public static void resetPrivateSession(Activity context, String sessionID) {
@@ -48,5 +58,10 @@ public class SharedPreferencesHelper {
     public static void setLockStatus(Context context, boolean lock) {
         SharedPreferences sharedPreferences = context.getSharedPreferences(context.getString(R.string.sharedPreferencesTag), MODE_PRIVATE);
         sharedPreferences.edit().putBoolean("LOCK", lock).apply();
+    }
+
+    public static boolean isLoggedIn(Context context) {
+        SharedPreferences sharedPreferences = context.getSharedPreferences(context.getString(R.string.sharedPreferencesTag), MODE_PRIVATE);
+        return sharedPreferences.getBoolean(Constants.logged_in, false);
     }
 }
